backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.3
beanie==1.26.0
pydantic[email]==2.7.0
python-multipart==0.0.9
python-dotenv==1.0.1
faker==25.0.0
pydicom==2.4.4
aiofiles==23.2.1
nibabel==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq>=0.9.0